refactor: 공지 리팩터링 (#114)
